Replace Search Console settings with more broad Google Integration settings
This changes the way Google-related settings are displayed. Instead of a
specific Search Console settings tab, this adds a Google Integration
tab. Inside this tab, the authentication-related UI elements are split
out into a standalone entry as the first setting, allow users to
interact with this to connect their Plausible and Google accounts. This
is not necessarily connected with Search Console. Then, the Search
Console settings panel just pertains to it, and not Google
authentication. A third panel is added for connecting Google Analytics
with the Plausible account.

The aim here is for a single point of Google authentication which is
needed for either one, or both, of Google Search Console or Google
Analytics integration.