Add search_console/analytics flags to GoogleAuth
This adds two boolean flags to a site's GoogleAuth configuration. These
are used to store whether a site wants integration with the search
console and/or historic google analytics data. An ecto migration is
added, enabling the search console boolean for existing GoogleAuths, as
currently the presence of google authentication implies integration with
Search Console.