Pass another dataset to main graph for rendering in red
This adds another entry to the JSON data returned via the main graph API
called `google_plot`, which is similar to `plot` in form but will be
completed with data downloaded via the Google Analytics reporting API.
Currently it simply returns the values from `plot` / 2. The data is
rendered in the main graph in red without fill, and without an indicator
for the present. Rationale: GA data will not continue to grow so there
is no projection forward, only backwards.