Fallback interval to localStorage saved value
This commit fixes a bug where the interval label was flickering when
changing the dashboard metric. This was caused because changing the
metric clears the `graphData` state property until the HTTP request
is finished.

This change fixes it by falling back interval to the value stored in
localStorage when `graphData` is falsy.