automatic CONVERSIONS -> PROPS navigation
In the Behaviours section, we'll navigate from Goal Conversions > Properties,
when a goal filter is applied by clicking on an entry in the Goal Conversions
list. This navigation does not happen for pageview goals or special goals (e.g.
'Outbound Link: Click', 'File Download', etc.)

The Behaviours component will remember that the navigation happened in order
to return the user to the previous view when the goal filter is removed.