Use `XHR` over `fetch` in all non-pageleave scripts
We're experiencing some ingestion-adjacent issues and want to confirm
whether it's caused by `fetch` changes.

We're not reverting it in `pageleave` script variants (used on
plausible.io and by a few other small clients) as it'd make scroll depth
and engagement time data gathered so far useless and in need to be
deleted.