Allow site transfer between different teams of the same user (#5192)
* Allow site transfer between different teams of the same user

* Fix typespec

* More type and branching fixes to satisfy dialyzer

* Alter when site transfers are included in sites listing

As now it's possible to transfer site between teams of the same user,
the site listing logic has to account for that. With this change, site
transfers are shown along sites under the condition that user is either
switched to personal sites view without a team or the team they
are currently switched to is different from the one the site belongs to
now.

* Make site danger zone available to admins

* Authorize admins to transfer sites on plug level

* Allow admin to initiate site transfer