APIv2: Fix autocomplete in Stats API playground (#5286)
Filters autocomplete only showed `"is", "segment"` under autocomplete
instead of all options. This is due to limitations in the vscode
autocomplete package which I don't fully understand but have to do with
partially overlapping `oneOf` definitions. As such the fix was to make
each `oneOf` element as independent of each other as possible.

Basecamp ref: http://localhost:8001/docs/stats-api-playground