Explicitly handle reverse order DateTimeRange
We're seeing warnings as follows:
```
 (plausible 0.0.1) lib/plausible/stats/legacy/legacy_query_builder.ex:32: Plausible.Stats.Legacy.QueryBuilder.from/4 |
 |  (plausible 0.0.1) lib/plausible/stats/query.ex:143: Plausible.Stats.Query.put_imported_opts/2 |
 |  (plausible 0.0.1) lib/plausible/stats/query.ex:163: Plausible.Stats.Query.get_imports_in_range/2 |
 |  (plausible 0.0.1) lib/plausible/imported.ex:98: Plausible.Imported.completed_imports_in_query_range/2 |
 |  (plausible 0.0.1) lib/plausible/stats/query.ex:65: Plausible.Stats.Query.date_range/2 |
 |  (elixir 1.17.3) lib/calendar/date.ex:111: Date.range/2 |
 |
warning: a negative range was inferred for Date.range/2, call Date.range/3 instead with -1 as third argument
```

As well as some Stats API queries with the date time range reversed.

This PR makes it explicit we support passing the range in reverse and
handles that without warnings. Along the way added some tests.

Ref: https://3.basecamp.com/5308029/buckets/36789884/card_tables/cards/8415153184