Fix up flaky tests
The suggest function sleeps for 500ms, so we need sufficient retries to handle
both the task execution time and LiveView message processing under low
resources. With this I couldn't reproduce anymore in repetead runs.

In the other case, funnel saved message may arrive after the test is
done and this is when Funnels are listed again.